+++ Sources/Sentry/SentryClient.m
@@ -3,11 +3,11 @@
 #import "SentryClient+Private.h"
 #import "SentryCrashDefaultMachineContextWrapper.h"
 #import "SentryCrashStackEntryMapper.h"
+#import "SentryDefaultTelemetryProcessorTransport.h"
 #import "SentryDefaultThreadInspector.h"
 #import "SentryDeviceContextKeys.h"
 #import "SentryEvent+Private.h"
 #import "SentryException.h"
-#import "SentryInstallation.h"
 #import "SentryInternalDefines.h"
 #import "SentryLogC.h"
 #import "SentryMechanism.h"
@@ -41,14 +41,15 @@ NS_ASSUME_NONNULL_BEGIN
 
 @protocol SentryEventContextEnricher;
 
-@interface SentryClientInternal () <SentryLogBufferDelegate>
+@interface SentryClientInternal ()
 
 @property (nonatomic, strong) SentryTransportAdapter *transportAdapter;
 @property (nonatomic, strong) SentryDebugImageProvider *debugImageProvider;
 @property (nonatomic, strong) id<SentryRandomProtocol> random;
 @property (nonatomic, strong) NSLocale *locale;
 @property (nonatomic, strong) NSTimeZone *timezone;
-@property (nonatomic, strong) SentryLogBuffer *logBuffer;
+@property (nonatomic, strong) id<SentryLogScopeApplier> logScopeApplier;
+@property (nonatomic, strong) id<SentryTelemetryProcessor> telemetryProcessor;
 @property (nonatomic, strong) id<SentryEventContextEnricher> eventContextEnricher;
 
 @end
@@ -122,9 +123,16 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
         self.attachmentProcessors = [[NSMutableArray alloc] init];
         self.eventContextEnricher = eventContextEnricher;
 
-        self.logBuffer = [[SentryLogBuffer alloc] initWithOptions:options
-                                                     dateProvider:dateProvider
-                                                         delegate:self];
+        self.telemetryProcessor = [SentryTelemetryProcessorFactory
+            getProcessorWithTransport:[[SentryDefaultTelemetryProcessorTransport alloc]
+                                          initWithTransportAdapter:transportAdapter]
+                         dependencies:SentryDependencyContainer.sharedInstance];
+
+        self.logScopeApplier =
+            [[SentryDefaultLogScopeApplier alloc] initWithEnvironment:options.environment
+                                                          releaseName:options.releaseName
+                                                   cacheDirectoryPath:options.cacheDirectoryPath
+                                                       sendDefaultPii:options.sendDefaultPii];
 
         // The SDK stores the installationID in a file. The first call requires file IO. To avoid
         // executing this on the main thread, we cache the installationID async here.
@@ -631,12 +639,12 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
 
 - (void)flush:(NSTimeInterval)timeout
 {
-    NSTimeInterval captureLogsDuration = [self.logBuffer captureLogs];
+    NSTimeInterval forwardingTelemetryDataDuration = [self.telemetryProcessor forwardTelemetryData];
     // Calculate remaining timeout for transport flush.
     // We subtract the time already spent capturing logs to respect the overall timeout.
     // If log capture took longer than the timeout, we use 0.0 which will still trigger
     // sending events but won't block waiting for completion.
-    NSTimeInterval remainingTimeout = fmax(0.0, timeout - captureLogsDuration);
+    NSTimeInterval remainingTimeout = fmax(0.0, timeout - forwardingTelemetryDataDuration);
     [self.transportAdapter flush:remainingTimeout];
 }
 
@@ -1096,22 +1104,28 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
 
 - (void)_swiftCaptureLog:(NSObject *)log withScope:(SentryScope *)scope
 {
-    if ([log isKindOfClass:[SentryLog class]]) {
-        [self.logBuffer addLog:(SentryLog *)log scope:scope];
+    if (self.options.enableLogs == NO) {
+        SENTRY_LOG_DEBUG(@"Dropping log, because the option enableLogs is false.");
+        return;
     }
-}
 
-- (void)captureLogs
-{
-    [self.logBuffer captureLogs];
-}
+    if ([log isKindOfClass:[SentryLog class]]) {
+        SentryLog *sentryLog = (SentryLog *)log;
+        SentryLog *enrichedLog = [self.logScopeApplier applyScope:scope toLog:sentryLog];
 
-- (void)captureLogsData:(NSData *)data with:(NSNumber *)itemCount
-{
-    [self captureData:data
-                 with:itemCount
-                 type:SentryEnvelopeItemTypes.log
-          contentType:@"application/vnd.sentry.items.log+json"];
+        // Call beforeSendLog callback if configured
+        if (self.options.beforeSendLog != nil) {
+            enrichedLog = self.options.beforeSendLog(enrichedLog);
+            if (enrichedLog == nil) {
+                SENTRY_LOG_DEBUG(@"Log dropped by beforeSendLog callback.");
+                [self recordLostEvent:kSentryDataCategoryLogItem
+                               reason:kSentryDiscardReasonBeforeSend];
+                return;
+            }
+        }
+
+        [self.telemetryProcessor addLog:enrichedLog];
+    }
 }
 
 - (void)captureMetricsData:(NSData *)data with:(NSNumber *)itemCount