+++ Sources/Swift/Tools/TelemetryBuffer/TelemetryItem.swift
@@ -1,4 +1,4 @@
-protocol TelemetryBufferItem: Encodable {
+protocol TelemetryItem: Encodable {
     var attributesDict: [String: SentryAttributeContent] { get set }
     var traceId: SentryId { get set }
 }