+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Changelog
 
-## Unreleased
+## 9.4.0
+
+### Breaking Changes
+
+- `Sentry.metrics.count(..)` does not support units, therefore the API was incorrectly defined. This breaking change applies to a method marked as experimental. (#7358)
+
+### Improvements
+
+- Record client report when dropping in `beforeSendLog` (#7379)
 
 ### Fixes
 
 - Fix mismatch of `in_foreground` app context (#7188) The app context `in_foreground` for handled and unhandled events was sometimes different. This is fixed now by aligning the implementation and adding a new `is_active` app context field.
 - Write reports on concurrent crashes (first wins) (#7340) Some concurrent crashes were incorrecly detected as re-crash on the same thread, and none has been written correcly as a consequence
+- Fix missing stack traces for `NSException` objects reported via `NSApplication.shared.reportException(_:)` on macOS (#7284)
 
 ## 9.3.0